requirements.txt
streamlit
torch
torchvision
ultralytics
usearch
clip-openai
opencv-python-headless
numpy
Pillow